FS-4566: adding designer multi input field edit subcomponent edit functionality
